+++ open_alchemy/column_factory/__init__.py
@@ -45,7 +45,11 @@ def column_factory(
     if type_ == "object":
         # Handle objects
         return object_ref.handle_object(
-            spec=spec, schemas=schemas, required=required, logical_name=logical_name
+            spec=spec,
+            schemas=schemas,
+            required=required,
+            logical_name=logical_name,
+            model_schema=model_schema,
         )
 
     if type_ == "array":